+++ CHANGELOG.md
@@ -42,6 +42,7 @@ CHANGELOG
 * @mmcc switched to using button elements for button components ([view](https://github.com/videojs/video.js/pull/2209))
 * @mmcc increased the size of the progress bar and handle on hover ([view](https://github.com/videojs/video.js/pull/2216))
 * @mmcc moved the fonts into their own repo ([view](https://github.com/videojs/video.js/pull/2223))
+* @mmcc deprecated the options() function and removed internal uses ([view](https://github.com/videojs/video.js/pull/2229))
 
 --------------------